Dimensionality Reduction - Linear Discriminant Analysis (LDA)
This project analyzes a wines dataset available to the public through the University of California Irvine Machine Learning repository. The dataset consists of 178 Italian wine samples, all from the same geographic region, but from 3 different cultivars. Each sample has 14 features: one is the cultivar's identifier, and the others are chemical attributes. The cultivar is a categorical variable that can take only 3 values: (Region) 1, (Region) 2, and (Region) 3. The chemical attributes are all numerical and continuous. 
In order to reduce the dimensionality, we have used Linear Discriminant Analysis  and used Logistic Regression to make the predictions on the test data set.